# docs/OpenAPIv3_swagger.yaml
components:
  schemas:
...
    domain.Chat:
      properties:
        created_at:
          type: string
        id:
          type: integer
        messages:
          items:
            $ref: '#/components/schemas/domain.Message'
          type: array
          uniqueItems: false
        metadata:
          description: Store ChatMetadata as JSON string
          type: string
        organization_id:
          type: integer
        tags:
          description: JSON array of tags as string
          type: string
        title:
          type: string
        updated_at:
          type: string
        user_id:
          description: Nullable for anonymous chats
          type: integer
      required:
      - created_at
      - id
      - messages
      - metadata
      - organization_id
      - tags
      - title
      - updated_at
      - user_id
      type: object
...
    handler.CreateChatRequest:
      properties:
        metadata:
          $ref: '#/components/schemas/domain.ChatMetadata'
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
        user_id:
          description: Optional, for anonymous chats
          type: integer
      required:
      - metadata
      - tags
      - title
      - user_id
      type: object
...
    handler.UpdateChatRequest:
      properties:
        metadata:
          $ref: '#/components/schemas/domain.ChatMetadata'
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        title:
          type: string
      required:
      - metadata
      - tags
      - title
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: x-organization-api-key
      type: apiKey
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: chatlogger-api-go+swagger@kjanat.com
    name: ChatLogger
    url: https://github.com/kjanat/chatlogger-api-go/issues
  description: API for logging and managing chat sessions.
  license:
    name: MIT License
    url: https://github.com/kjanat/chatlogger-api-go/blob/master/LICENSE
  termsOfService: https://github.com/kjanat/chatlogger-api-go#terms-of-service
  title: ChatLogger API (Go)
  version: ""
openapi: 3.1.0
paths:
  /v1/analytics/messages:
    get:
      description: Retrieves aggregated statistics about messages within a specified
        date range for the user's organization.
      parameters:
      - description: Start date (RFC3339 format, e.g., 2023-01-01T00:00:00Z). Defaults
          to 30 days ago.
        in: query
        name: start
        schema:
          type: string
      - description: End date (RFC3339 format, e.g., 2023-01-31T23:59:59Z). Defaults
          to now.
        in: query
        name: end
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Aggregated message statistics
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid date format
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get message statistics
      security:
      - BearerAuth: []
      summary: Get Message Statistics
      tags:
      - Analytics
  /v1/chats:
    get:
      description: Retrieves a paginated list of chat sessions for the user's organization.
      parameters:
      - description: Number of chats per page
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      - description: Offset for pagination
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/domain.Chat'
                type: array
          description: List of chats
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list chats
      security:
      - BearerAuth: []
      summary: List Chats
      tags:
      - Chats
    post:
      description: Creates a new chat session for the organization. Can be called
        via Dashboard (JWT) or Public API (API Key).
      parameters:
      - description: Organization Slug (Required for Public API)
        in: path
        name: slug
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.CreateChatRequest'
        description: Chat Details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'message: Chat created successfully, chat_id: uint64'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT or API Key invalid/missing, or Org ID not
            found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden (API Key doesn't match slug)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create chat or process tags/metadata
      security:
      - ApiKeyAuth: []
        BearerAuth: []
      summary: Create Chat
      tags:
      - Chats
  /v1/chats/{chatID}:
    delete:
      description: Deletes a chat session and its associated messages.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat deleted successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get or delete chat
      security:
      - BearerAuth: []
      summary: Delete Chat
      tags:
      - Chats
    get:
      description: Retrieves details for a specific chat session, optionally including
        messages.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      - description: Include messages in the response
        in: query
        name: include_messages
        schema:
          default: false
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.GetChatResponse'
          description: Chat details
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get chat or messages
      security:
      - BearerAuth: []
      summary: Get Chat by ID
      tags:
      - Chats
    patch:
      description: Updates the title, tags, or metadata of an existing chat session.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.UpdateChatRequest'
        description: Fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handler.GetChatResponse'
          description: Updated chat details
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID or request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get or update chat, or process tags/metadata
      security:
      - BearerAuth: []
      summary: Update Chat
      tags:
      - Chats
  /v1/chats/{chatID}/messages:
    get:
      description: Retrieves all messages associated with a specific chat session.
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/handler.GetMessageResponse'
                type: array
          description: List of messages with parsed metadata
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Permission denied (Chat doesn't belong to user's org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get chat or messages
      security:
      - BearerAuth: []
      summary: Get Chat Messages
      tags:
      - Messages
  /v1/exports:
    get:
      description: Retrieves a paginated list of asynchronous export jobs for the
        user's organization.
      parameters:
      - description: Number of exports per page
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      - description: Offset for pagination
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/domain.Export'
                type: array
          description: List of export jobs
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to fetch exports
      security:
      - BearerAuth: []
      summary: List Export Jobs
      tags:
      - Exports
    post:
      description: Initiates an asynchronous job to export chat data (chats, messages,
        or all) in JSON or CSV format for the user's organization.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ExportRequest'
        description: Export parameters (format, type)
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'export_id: uint64, status: domain.ExportStatus, message: string'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data (format/type)
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing, Org or User ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create export job
      security:
      - BearerAuth: []
      summary: Create Asynchronous Export Job
      tags:
      - Exports
  /v1/exports/{id}:
    get:
      description: Retrieves the status and details of a specific asynchronous export
        job.
      parameters:
      - description: Export Job ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain.Export'
          description: Export job details
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid export ID
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Export not found or doesn't belong to user's org
      security:
      - BearerAuth: []
      summary: Get Export Job Status
      tags:
      - Exports
  /v1/exports/{id}/download:
    get:
      description: Downloads the file generated by a completed asynchronous export
        job.
      parameters:
      - description: Export Job ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Export file (JSON or CSV)
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid export ID or export not ready
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Export not found or doesn't belong to user's org
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Export file path not found or file missing on disk
      security:
      - BearerAuth: []
      summary: Download Export File
      tags:
      - Exports
  /v1/exports/sync:
    post:
      description: Immediately generates and returns an export file (JSON or CSV)
        containing chat data. Use async export for large datasets.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ExportRequest'
        description: Export parameters (format, type)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: file
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: Export file (JSON or CSV)
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data (format/type)
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or Org ID not found)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to retrieve data or generate export
      security:
      - BearerAuth: []
      summary: Create Synchronous Export (Legacy)
      tags:
      - Exports (Legacy)
  /v1/orgs/{slug}/chats:
    post:
      description: Creates a new chat session for the organization. Can be called
        via Dashboard (JWT) or Public API (API Key).
      parameters:
      - description: Organization Slug (Required for Public API)
        in: path
        name: slug
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.CreateChatRequest'
        description: Chat Details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'message: Chat created successfully, chat_id: uint64'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT or API Key invalid/missing, or Org ID not
            found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden (API Key doesn't match slug)
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create chat or process tags/metadata
      security:
      - ApiKeyAuth: []
        BearerAuth: []
      summary: Create Chat
      tags:
      - Chats
  /v1/orgs/{slug}/chats/{chatID}/messages:
    post:
      description: Adds a new message to an existing chat session. Can be called via
        Public API (API Key).
      parameters:
      - description: Organization Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Chat ID
        in: path
        name: chatID
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.CreateMessageRequest'
        description: Message Details (role, content, metadata)
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: 'message: Message created successfully, message_id: uint64'
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid chat ID or request data (role, content, metadata validation)
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (API Key invalid/missing or Org ID not found)
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden (API Key doesn't match slug or chat doesn't belong
            to org)
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Chat not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to get chat or create message
      security:
      - ApiKeyAuth: []
      summary: Create Message
      tags:
      - Messages
  /v1/orgs/me/apikeys:
    get:
      description: Lists all API keys for the organization associated with the authenticated
        user.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/domain.APIKey'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized or Org ID not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list API keys
      security:
      - BearerAuth: []
      summary: List API Keys
      tags:
      - API Keys (Admin)
    post:
      description: Generates a new API key for the organization associated with the
        authenticated user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.GenerateKeyRequest'
        description: API Key Label
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: API key generated successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized or Org ID not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to generate API key
      security:
      - BearerAuth: []
      summary: Generate API Key
      tags:
      - API Keys (Admin)
...
  /v1/users/me/password:
    post:
      description: Allows the currently authenticated user to change their password.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ChangePasswordRequest'
        description: Current and new password
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Password changed successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request data or password change failed (e.g., wrong
            current password)
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized (JWT invalid/missing or User ID not found)
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Users
...